Fix Windows portability: use ptrdiff_t in operator- test
Agent-Logs-Url: https://github.com/GregorGullwi/FlashCpp/sessions/997f5b02-8583-49dd-a0bd-30dc5741d16b

+++ tests/test_member_operator_template_pointer_self_ref_minus_ret0.cpp
@@ -1,8 +1,10 @@
+#include <stddef.h>
+
 template<typename T>
 struct Iter {
 	T* ptr;
 
-	long operator-(const Iter& other) const;
+	ptrdiff_t operator-(const Iter& other) const;
 };
 
 int main() {
